32b/64b compatibility fixes
In a few places, types like int or uint32_t were used when the intent
was to use a type as wide as a machine register or a pointer.
I replaced these mostly with uintptr_t.

+++ kern/arch/i686/process.c
@@ -9,7 +9,7 @@
 #include <stdio.h>
 
 void proc_init_trapframe(trapframe_t *tf, uint32_t vcoreid,
-                         uint32_t entryp, uint32_t stack_top)
+                         uintptr_t entryp, uintptr_t stack_top)
 {
 	memset(tf,0,sizeof(*tf));
